Testing Interpreter
We are using Pythong 3.14.0 FYI
Just wanted to make sure that the interpreter was working.

Originally thought I had it installed but now I know if you do have it installed at least on your C:/ drive VSC will automatically bring it up when youre looking to link the interpreter. Exceptions are likely when its in a different drive than the OS E.G. external SSD.