Create LightGlueSource.py
Was retesting to see if I had the correct Python Version installed and the interpreter was pathed correctly to VSC.

Turns out it was and on top of that the issue was that when calling the following "python -m venv venv" it wasnt working and when even adressing a specific file path it wasnt working due to the fact that I was not using ampersand in the terminal.

I was unaware that the terminal in VSC by default is using PowerShell thus you need to state "& 'C:\\ your path' -m venv [name]" to create the virtual environment.