Rollback to Py312
Learned that torch depends on 312 in order to run properly and I made the environment in 314.

Deleted VEnv #1 to make a 2nd one
Made it with 312
Now installing torch for lightglue.

Remembered that I can make a .gitignore file learning its common practice to not commit your VEnv ever... [VEnv = Virtual Environment]