Organization
Was failing to utilize Git to update the changes I made and why, I changed the labels of the images to make it easier to organize and figure out what is happening and where.

Made a run file finally as the imports test was successful after adressing PATH issues with Python, I just didnt clear the cache to adress the interpreter path correctly.

Remade the VEnv, and hence all these organizational changes. Now working on using correct definitons [keys] in the lightglue run file