Update run_lightglue_match.py
This update adresses the fact that we didnt pull the compiled data from the GPU, it was stuck on memory and we cant call it in its current state to be utilized to develop an image.

Must be converted to CPU which it then can be processed, through numpy to adress the pixel values for image comparison properly. However these will be converted from float to int as we dont have subpixel precision in image processing... yet??? not sure on that one.

After we convert the comparison vector end point [thats what Imma call them for now as I havent read documentation yet]

We will be able to have images that show the comparisons... hopefully?